Add the cypress-wait-until plugin (#1625)
A lot of StackOverflow users had some difficulties in implementing a recursive promise with Cypress just to repeatedly check for something to happen (see two of the various questions about the topic: How can i wait for each element in a list to update to a certain text? And How do I wait until a cookie is set?). 
We wrote this plugin to file this kind of issues