📝 Include a much simpler TypeScript example (#1745)
* 📝 Include much simple TypeScript example

There was an overload of information in these docs (and example recipes)
and it took me several hours to set-up something that was relatively
simple if it was communicated better. Feel free to fork the repo I
put in the links. Please put something this simple in these docs.

* 👌 Changes according to PR reviewer

Changed typescript to TypeScript