Make JSON Web tokens section match rest of page format
clicking on the The "JSON web tokens" hash took a user to the github page instead of the hash fragment url. JSON web tokens did not have a description section on the page either. 

I fixed both of these issues to make JSON web tokens match the formatting of the rest of the page and copied the description info from the github page.