Suggest cypress-multi-reporters
mocha-multi-reporters is broken for mocha 6+, which is what users will get if they run the suggested command:

```
npm install --save-dev mocha mocha-multi-reporters mocha-junit-reporter
```

However, there's a maintained fork of mocha-multi-reporters which supports mocha v6+, cypress-multi-reporters.  This commit changes the docs to suggest that newer package instead of the broken one.