Added example for serving instrumenting code
The documentation explains how to setup the babel-plugin-istanbul plugin, but doesn't provide any examples or guidance for how to serve the application with the instrumented code without an intermediate folder, once the plugin is setup. 

I've also separated the Instrumenting code section into the two parts, which should make it easier to read. 

Potential limitation of this specific example is that it assumes the application was build using create-react-app.