Clarification on return value
There are many situations when beginners might stumble over this. Here is one example:

cy.get('#domElemWithTextContent').should('have.length.of.at.least', 5)      <= DOES NOT WORK!!!   this checks for the length of the array not the length of the text!
cy.get('#domElemWithTextContent').invoke('text').should('have.length.of.at.least', 5)         <= Only this works.