CLOUD-155: Add GitLab Self-managed to Dashboard GitLab Integration docs
This change adds minimal edits to the Dashboard -> GitLab Integration page to briefly explain pitfalls & additional steps specific to the new GitLab Self-managed integration.

Please note that this change is not ready to be merged as there are outstanding issues to be resolved - whilst we have confirmed the renaming of 'GitLab for Enterprise' to 'GitLab Self-managed', the required Dashboard changes have yet to land & may yet require further changes.

Furthermore, the Dashboard's integration configuration areas are currently undergoing redesign, so consider this a precursor to a more thorough rewrite of the integration docs that will address their existing shortcomings.