CLOUD-295: Add analytics aspect to "Why Cypress -> Our mission"
This, possibly contentious, change adds language to the Our Mission section that emphasizes the analysis of test results as a key differentiator of Cypress in general, and by inference the dashboard specifically. I've tried to suggest a more general mission of leveraging test metadata to achieve a higher level of software refinement, hopefully it hasn't just degenerated into word salad.

Note also that I have added the new content to one of the existing paragraphs rather than create a new one. This is purely a subjective aesthetic preference on my part, I felt that this content could be considered an addendum to or refinement of the language in that para.