position is not an option but an argument
In the click-documentation, `position` is listed as an argument, not a s an option. Unfortunately i can't judge whether one or the other works as this seems to be flaky in our case.
I just wanted to do a PR instead of an issue in order to be more constructive in the case that i'm right but maybe i'm wrong here for some reason?