doc(url.mdx): update description
The description on this one is quite confusing. It stated that the command `cy.url()` yields the url as a string which implies that it directly returns the current url as a string. Whereas in reality it returns the Promise like object that actually resolves the url as a string. This confused me and got me stuck for a bit until I found the correct way of doing it from Stackoverflow which is:

```js
cy.url().then(url => {
  // the actual url string
});
```