Initial changes for search engine SERP adjustments
Change how the docs pages are rendered in search results.  Doesn't impact SEO, but impacts how those existing search results are displayed to the user.  SEO will be done separately.
This is phase 1 of the changes - adjusting 10 pairs of title and description meta tags to confirm the desired behavior before implementing the rest.