chore: title of some pages change for SEO (#5728)
* Initial changes for search engine SERP adjustments

Change how the docs pages are rendered in search results.  Doesn't impact SEO, but impacts how those existing search results are displayed to the user.  SEO will be done separately.
This is phase 1 of the changes - adjusting 10 pairs of title and description meta tags to confirm the desired behavior before implementing the rest.

* Fix - quote meta tags to avoid special symbols

single quote all the meta tags as strings in order to comply with YAML spec.  Docusaurus did not parse out the colon symbols and interpreted those as a special tag.

* fix: make sure we can use the title in markdown

* feat: make sure all pages have a one line visual title

---------

Co-authored-by: djconsult <133032924+danjohansenconsulting@users.noreply.github.com>