Framework deployment for new features
Reorganized the docs structure so that analytics is now a submenu.  The entry page describes the types of analytics and API available with links to each type.  The original /analytics is now one of the sub-pages, called project-analytics.

The API framing page is under the integrations area of the docs.

Both API and enterprise reporting have info blocks to explain that the features are "coming soon."  We will deploy a secondary deployment this week to fill in the details (it will go GA on Monday so we don't want to deploy it too early).  But for now we want the docs pages published so that Cloud can link to the correct section of the docs.