Teams in Cloud - Draft 1 for release (#6125)
* Draft 1 for release

Initial description of the new functionality provided in Cloud.  This is currently in closed beta but will go GA to all organizations this week.

We will have follow-up content with more detail, especially regarding the new user roles and user page filters, in subsequent PRs.

* Revisions for clarity

Added alt text to the images and provide common use cases for organizations wanting to use this Teams feature