Update Happo plugin description and link (#6322)
We have merged the happo-cypress package into the main happo package, so
we need to update this link.

The repo can be found at https://github.com/happo/happo but since it
handles more than Cypress I thought it might make the most sense to link
to the Cypress integration documentation from this page.

I am also interested in getting the badge upgraded from community to
verified, but I assume that someone at Cypress will need to handle that. Let me know if I should open a PR for this!

Co-authored-by: Jennifer Shehane <jennifer@cypress.io>